requirements.txt
runlocal_tools==0.0.2
huggingface_hub
coremltools
torch
torchvision
onnx
onnxruntime
onnxscript
pillow